test(SPV-1094): create fixtures and change file structure
