+++ jest.config.ts
@@ -154,7 +154,7 @@ export default {
   // ],
 
   // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
-  testPathIgnorePatterns: ['/node_modules/', '/regression_tests/'],
+  testPathIgnorePatterns: ['/node_modules/', '/regression_tests/', '__tests__/fixtures/'],
 
   // The regexp pattern or array of patterns that Jest uses to detect test files
   // testRegex: [],
